organized cached types and specs to match
