+++ spec/cacheable/types/attribute_cache_spec.rb
@@ -1,0 +1,60 @@
+require 'spec_helper'
+
+describe Cacheable do
+  let(:cache) { Rails.cache }
+  let(:user)  { User.create(:login => 'flyerhzm') }
+
+  before :all do
+    @post1 = user.posts.create(:title => 'post1')
+    @post2 = user.posts.create(:title => 'post2')
+  end
+
+  before :each do
+    cache.clear
+    user.reload
+  end
+
+  context "with_attribute" do
+    it "should not cache User.find_by_login" do
+      Rails.cache.read("users/attribute/login/flyerhzm").should be_nil
+    end
+
+    it "should cache by User.find_by_login" do
+      User.find_cached_by_login("flyerhzm").should == user
+      Rails.cache.read("users/attribute/login/flyerhzm").should == user
+    end
+
+    it "should get cached by User.find_by_login multiple times" do
+      User.find_cached_by_login("flyerhzm")
+      User.find_cached_by_login("flyerhzm").should == user
+    end
+
+    it "should escape whitespace" do
+      new_user = User.create(:login => "user space")
+      User.find_cached_by_login("user space").should == new_user
+    end
+
+    it "should handle fixed numbers" do
+      Post.find_cached_by_user_id(user.id).should == @post1
+      Rails.cache.read("posts/attribute/user_id/#{user.id}").should == @post1
+    end
+
+    context "find_all" do
+      it "should not cache Post.find_all_by_user_id" do
+        Rails.cache.read("posts/attribute/user_id/all/#{user.id}").should be_nil
+      end
+
+      it "should cache by Post.find_cached_all_by_user_id" do
+        Post.find_cached_all_by_user_id(user.id).should == [@post1, @post2]
+        Rails.cache.read("posts/attribute/user_id/all/#{user.id}").should == [@post1, @post2]
+      end
+
+      it "should get cached by Post.find_cached_all_by_user_id multiple times" do
+        Post.find_cached_all_by_user_id(user.id)
+        Post.find_cached_all_by_user_id(user.id).should == [@post1, @post2]
+      end
+
+    end
+  end
+
+end
